+++ examples/01_encodings.py
@@ -39,8 +39,8 @@ We study the case of predicting wages using the
 """
 
 ###############################################################################
-# Easily encoding a dataframe
-# ---------------------------
+# Easy learning on a dataframe
+# ----------------------------
 #
 # Let's first retrieve the dataset, using one of the downloaders from the
 # :mod:`skrub.datasets` module. As all the downloaders,
@@ -57,22 +57,42 @@ employees, salaries = dataset.X, dataset.y
 employees
 
 ###############################################################################
-salaries
+# Most machine-learning algorithms work with arrays of numbers. The
+# challenge here is that the ``employees`` dataframe is a heterogeneous
+# set of columns: some are numerical (``'year_first_hired'``), some dates
+# (``'date_first_hired'``), some have a few categorical entries
+# (``'gender'``), some many (``'employee_position_title'``). Therefore
+# our table needs to be "vectorized": processed to extract numeric
+# features.
+#
+# ``skrub`` provides an easy way to build a simple but reliable
+# machine-learning model which includes this step, working well on most
+# tabular data.
 
-###############################################################################
-# We observe diverse columns in the ``employees`` dataframe:
-#   - numeric (``'year_first_hired'``)
-#   - dates (``'date_first_hired'``)
-#   - low-cardinality categorical (``'gender'``, ``'department'``,
-#     ``'department_name'``, ``'assignment_category'``)
-#   - high-cardinality categorical (``'employee_position_title'``, ``'division'``).
+from sklearn.model_selection import cross_validate
+
+from skrub import tabular_learner
+
+model = tabular_learner("regressor")
+results = cross_validate(model, employees, salaries)
+results["test_score"]
+
+# %%
+# The estimator returned by :obj:`tabular_learner` combines 2 steps:
 #
-# Most machine-learning algorithms work with arrays of numbers. Therefore our complex,
-# heterogeneous table needs to be processed to extract numeric features. Transforming a
-# complex real-world object such as a date into a vector of numeric features —more
-# adequate for machine learning— is often called *vectorizing* it.
+# - a |TableVectorizer| to preprocess the dataframe and vectorize the features
+# - a supervised learner (by default a |HGBR|)
+model
+
+# %%
+# In the rest of this example, we focus on the first step and explore the
+# capabilities of skrub's |TableVectorizer|.
 #
-# We can easily do this using skrub's |TableVectorizer|.
+# |
+
+# %%
+# More details on encoding tabular data
+# -------------------------------------
 
 from skrub import TableVectorizer
 
@@ -240,12 +260,14 @@ print(f"mean fit time: {np.mean(results['fit_time']):.3f} seconds")
 # of particular character n-grams (more details are provided in its documentation).
 # Therefore it can be a faster and very effective alternative, when the supervised
 # learner is built on top of decision trees, which is the case for the |HGBR|.
+#
+# The resulting pipeline is identical to the one produced by default by
+# :obj:`tabular_learner`.
 
 from skrub import MinHashEncoder, ToCategorical
 
 vectorizer = TableVectorizer(
-    low_cardinality_transformer=ToCategorical(),
-    high_cardinality_transformer=MinHashEncoder(),
+    low_cardinality=ToCategorical(), high_cardinality=MinHashEncoder()
 )
 pipeline = make_pipeline(
     vectorizer, HistGradientBoostingRegressor(categorical_features="from_dtype")