+++ skrub/tests/test_gap_encoder.py
@@ -224,9 +224,11 @@ def test_missing_values(df_module):
     observations = df_module.make_column("", observations)
     enc = GapEncoder(n_components=3)
     enc.fit_transform(observations)
+    enc.transform(observations)
     enc.partial_fit(observations)
     # non-regression for https://github.com/skrub-data/skrub/issues/921
-    assert sbd.is_null(observations)[2]
+    for null_idx in 2, 4, 5:
+        assert sbd.is_null(observations)[null_idx]
 
 
 def test_check_fitted_gap_encoder(df_module):
@@ -289,3 +291,15 @@ def test_output_pandas_index():
     s_test = pd.Series("one two two".split(), name="", index=[-11, 200, 32])
     out = gap.transform(s_test)
     assert out.index.tolist() == [-11, 200, 32]
+
+
+def test_duplicate_topics(df_module):
+    s = df_module.make_column("s", ["one", "two"] * 10)
+    out = GapEncoder(n_components=3, random_state=0).fit_transform(s)
+    assert sbd.column_names(out) == ["s: one, two", "s: two, one", "s: one, two (2)"]
+
+
+def test_empty_column_name(df_module):
+    s = df_module.make_column("", ["one", "two"] * 10)
+    out = GapEncoder(n_components=3, random_state=0).fit_transform(s)
+    assert sbd.column_names(out) == ["one, two", "two, one", "one, two (2)"]